extend read-only region 1 back from start.
Otherwise text written direclty before the read-only
region is overwritten on copy-back.

I'm not really sure if i'd consider it an "off by 1" error
though. Becuase now it feels as if the locking is 1 character
"too strong" preventing legit edits in front of the region.
However for now, the extra safety of not having legit text
deleted is prefered.